+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 ========
 Extension "Locations" for Contao Open Source CMS
 ========
+0.6.0 - 2018-01-09
+- Add categories to markers
+
+0.5.0 - 2018-01-07
+- Add tl_content as child table of tl_wem_location, so we can have a subpage for each item
+
+0.4.0 - 2018-01-04
+- Rework the way maps, options and markers are handle
 
 0.3.0 - 2018-09-24
 - Add Leaflet library, thanks to Ollie